Add WSGI Authorization pass
By default, when using WSGI through httpd, WSGI will consume the
'Authorization' header before passing the headers onto the application.

This was preventing the application itself from processing the
authenication.  There is a WSGI setting however that allows the header
to pass onto the application, so had to set that via the ebextensions.