Refactor setup to to fix test cases
With the addition of the ssm store, the setup was no longer purely
local, which was causing issues with the tests.

The ssm calls are now only done via the functions that need them and
then stored into the configuration, which appears to be global.

Also had to encode the form data to utf-8 to work with bcrypt.